Add `schema.sql` for creating database tables
This new `schema.sql` script should be used for setting up the database
tables this app will use. Note that if you need to run this script (say,
if you want to run the app locally and need to configure your database),
you should first create the `calculator_service` database.

I have created my own `setup.sql` script which does this. This file will
not be committed as it also creates a user for interacting with the
database, and I don't want to expose the user's password.