Add versioning for API endpoints
I've introduced this code to add a version to my current set of API
routes. If I ever want to introduce an API v2, I can set up routes
prefixed with `v2`.

I currently only have a login route, so this seemed like a good time to
introduce this change.