Add decorator for auth on protected routes
This new decorator will be applied to routes which should be protected
from unauthorized use. It searches for a bearer token in the request and
validates it before processing the request.

For now I'm using the user JWT which is generated when a user is logged
in. If I run into roadblocks with reusing the login system's
authorization I can always create a system for long-lived tokens.