Add implementation for operation creation route
To handle creating a new operation in the database, I've added a new
method to the DBService class. This new method will handle inserting new
records.

Unit tests pass, but I'd like to add a few more unit tests to check
responses when bad data is sent in the request.