Add script for generating an admin API key
This script also stores the key in the database. Currently the script is
only set up to work with a locally running database; I'll come back
later and update it to allow a user to connect to the production
database. When generating a production admin key, the user will be
prompted for database credentials in order to connect.