Implement route for running a calculation
This route definition already has a fair bit of error handling, but I'm
sure I'll be adding much more in the near future.

Note that calculation logic is deferred to the "CalculatorService"
class. An administrator can use the "POST /operations" endpoint to
easily add a new operation to the database, but all operation logic will
also need to be defined in this class. If a requested operation exists
in the database but the logic isn't defined, a "NotImplementedError"
exception is raised.

I've only defined logic for "addition", so I'll now go through and add
some logic for the other operation types.