Implement route to soft-delete a calc record
This route will be used by an "admin" user to run a soft-delete on a
recorded calculation.

Since I want to enable an admin to delete any previous calculation, and
subsequent calculations base their "user_balance" value on the previous
update, this requires a cascade-style update to all of the user's other
recorded calculations to correct the user balance.